# Make linear reads optional